Added support to the OSM v8
The Multi-SFC was using an old client API to interact with the Open
Source MANO platform. This API was included in the project and required
code updates every single time the OSM bumped into a new version.

This change adds the support of the OSM version 8 by employing the osmclient
API provided by the OSM developers. In this way, when a new feature is
added to the OSM, it is only required in most of the cases, updating the
python osmclient library.

Signed-off-by: Alexandre Huff <alexandrehuff@gmail.com>