Added support to federation
This change adds support to federation in Multi-SFC using the SAML
protocol. Several changes have been made to fit the current
implementation of the core module as well as in the Tacker and OpenStack
drivers. OSM driver does not provide support to federation due to
lack of federation support in the OSM orchestrator itself, so far.