Added federated authentication to the Multi-SFC API
This change implements the support of federated authentication and
authorization in the Multi-SFC API. The Multi-SFC client was modified to
support the ECP message flow between the SP (i.e. Multi-SFC) and any IdP
that implements the SAML2 protocol.
Tests have been done using the SAMLtest.id IdP.

Several configuration files were added to allow building a Multi-SFC
container using the docker-compose. Federation support is configured in
the container automatically. A Dockerfile allows to choose whether or not
the federation support is enabled.